process live games

+++ model/src/main/java/com/eb/schedule/model/MatchStatus.java
@@ -4,19 +4,19 @@ package com.eb.schedule.model;
  * Created by Egor on 12.02.2016.
  */
 public enum MatchStatus {
-    SCHEDULED(0),
-    LIVE(1),
-    FINISHED(2);
+    SCHEDULED((byte)0),
+    LIVE((byte)1),
+    FINISHED((byte)2);
 
 
     public final byte status;
 
-    MatchStatus(int status) {
-        this.status = (byte) status;
+    MatchStatus(byte status) {
+        this.status = status;
     }
 
 
-    public static MatchStatus fromValue(int status){
+    public static MatchStatus fromValue(byte status){
         for(MatchStatus matchStatus : MatchStatus.values()){
             if( matchStatus.status == status){
                 return matchStatus;
